+++ src/misc/ft_2d_arrlen.c
@@ -6,15 +6,17 @@
 /*   By: rubennijhuis <rubennijhuis@student.coda      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/24 19:09:04 by rubennijhui   #+#    #+#                 */
-/*   Updated: 2022/05/07 12:09:55 by rubennijhui   ########   odam.nl         */
+/*   Updated: 2022/06/30 20:39:50 by rubennijhui   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-uint32_t	ft_2d_arrlen(char **arr)
+#include <stddef.h>
+
+size_t	ft_2d_arrlen(char **arr)
 {
-	uint32_t	i;
+	size_t	i;
 
 	i = 0;
 	while (arr[i] != NULL)
